Start using a mock timeline controller for the UserSession flows
